requirements.txt
streamlit
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai
langchain-chroma
langchain-huggingface
sentence-transformers
chromadb
python-dotenv
tiktoken
pypdf
docx2txt
jq
pandas
